backend/requirements.txt
fastapi
uvicorn
pytest
psycopg2-binary
pydantic_settings
SQLAlchemy
#mysql-connector-python
#pytest to test connection with db ( the ip of container of db is not localhost because the default network is bridge)
